requirements.txt
beautifulsoup4==4.11.1
Flask==1.0.2
jieba==0.42.1
marshmallow==3.18.0
py_cpuinfo==8.0.0
requests==2.21.0
SQLAlchemy==1.4.41
